See #ANT-958 Move ModelSelectionStrategy to simulation directory
