+++ src/andromede/model/__init__.py
@@ -12,15 +12,7 @@
 
 from .common import ProblemContext, ValueType
 from .constraint import Constraint
-from .model import (
-    InvestmentProblemStrategy,
-    MergedProblemStrategy,
-    Model,
-    ModelPort,
-    ModelSelectionStrategy,
-    OperationalProblemStrategy,
-    model,
-)
+from .model import Model, ModelPort, model
 from .parameter import Parameter, float_parameter, int_parameter
 from .port import PortField, PortType
 from .variable import Variable, float_variable, int_variable
